Remove namespace from values
Helm has a --namespace argument for a reason.  Using a non-standard mechanism
like putting the namespace in the values file "breaks" tools like argo and
flux, and confuses people who expect the helm .Release.Namespace value to be
respected. This change means future expansion is easier, since there is no need
to remember to add an explicit namespace to every single resource. :)